% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{basin_size}
\alias{basin_size}
\title{Basin size}
\format{
A data frame with 7 observations on the following 5 variables.
\describe{ \item{ID_BASIN_SIZE}{a numeric vector}
\item{BASIN_SIZE_CODE}{a character vector}
\item{BASIN_SIZE_DESC}{a character vector}
\item{BASIN_SIZE_RANGE_MIN}{a character vector}
\item{BASIN_SIZE_RANGE_MAX}{a character vector} }
}
\description{
Basin size
}
\examples{

data(basin_size)
## maybe str(basin_size) ; plot(basin_size) ...

}
\keyword{datasets}
