% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{charcoal_units}
\alias{charcoal_units}
\title{Charcoal units}
\format{
A data frame with 169 observations on the following 4 variables.
\describe{ \item{unit}{a factor with levels \code{\%125} \code{\%DWT}
\code{\%TOF} \code{\%WTO} \code{100P} \code{1012} \code{1220} \code{1422}
\code{1424} \code{3869} \code{5015} \code{5075} \code{6914} \code{7510}
\code{7738} \code{7PRA} \code{8POS} \code{ABSC} \code{ARCC} \code{AREA}
\code{ARPG} \code{BCCT} \code{C004} \code{C010} \code{C01K} \code{C040}
\code{C100} \code{C120} \code{C125} \code{C140} \code{C150} \code{C180}
\code{C250} \code{C65X} \code{CARE} \code{CCMY} \code{CG2M} \code{CHCO}
\code{CHRX} \code{CM10} \code{CMGR} \code{CMML} \code{COML} \code{CP20}
\code{CPAR} \code{CPRA} \code{DC13} \code{FRAG} \code{G05I} \code{G10I}
\code{G250} \code{GCHG} \code{GCMY} \code{GRG1} \code{GRGR} \code{GRPC}
\code{GT05} \code{GT10} \code{GT12} \code{GT15} \code{GT18} \code{GT1M}
\code{GT24} \code{GT25} \code{GT2X} \code{GT30} \code{GT38} \code{GT50}
\code{GT53} \code{GT5C} \code{LT12} \code{LT13} \code{LT18} \code{LT1M}
\code{LT25} \code{LT40} \code{M2GR} \code{M2ML} \code{MCM3} \code{MM2C}
\code{MM2G} \code{MMDW} \code{MMML} \code{MT10} \code{MULT} \code{NOTK}
\code{ORDI} \code{OTHE} \code{PCM1} \code{PCMY} \code{PHYT} \code{PP25}
\code{PP51} \code{PP52} \code{PPML} \code{PPOL} \code{PRAB} \code{PROB}
\code{PTC0} \code{PTC2} \code{PTC3} \code{PTCT} \code{PX25} \code{SQC0}
\code{SQCU} \code{SQCY} \code{SQG1} \code{SQG5} \code{SQL5} \code{TOCA}
\code{TOM1} \code{TOM2} \code{TOM3} \code{TOM4} \code{TOM5} \code{TOM6}
\code{TOM7} \code{TOM8} \code{TOM9} \code{TOMX} \code{TOTA} \code{UCMY}
\code{V025} \code{V100} \code{V125} \code{V255} \code{V501} \code{X01K}
\code{X105} \code{X106} \code{X120} \code{X125} \code{X150} \code{X15G}
\code{X160} \code{X180} \code{X18C} \code{X20P} \code{X250} \code{X25P}
\code{X310} \code{X37P} \code{X459} \code{X500} \code{X50U} \code{X512}
\code{X515} \code{X520} \code{X53C} \code{X550} \code{X55P} \code{X55U}
\code{XA1K} \code{XARE} \code{XARP} \code{XARS} \code{XC25} \code{XCM3}
\code{XCOP} \code{XFML} \code{XFRG} \code{XFRP} \code{XFRS} \code{XIMG}
\code{XPEA} \code{XPEE} \code{XPIX} \code{XRIF}} \item{qtype}{a
factor with levels \code{C0P0} \code{CONC} \code{INFL} \code{NOTK}
\code{OTHE} \code{SOIL}} \item{type}{a character vector}
\item{code}{a numeric vector} }
}
\description{
Charcoal units used in GCD
}
\examples{

data(charcoal_methods)
## maybe str(charcoal_methods) ; plot(charcoal_methods) ...

}
\keyword{datasets}
