% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{date}
\alias{date}
\title{Date table}
\format{A data frame with 4872 observations on the following 8 variables.
\describe{ \item{ID_SITE}{a numeric vector}
\item{DATE_LAB_NUMBER}{a factor with many levels}
\item{DEPTH}{a numeric vector} \item{TICKNESS}{a numeric
vector} \item{AGE}{a numeric vector} \item{ERROR}{a numeric
vector} \item{ID_DATE_TYPE}{a numeric vector}
\item{ID_MAT_DATED}{a factor with levels} }
obtained ~~}
\description{
Table with the dating information contained in GCD
}
\examples{

data(date)
## maybe str(date) ; plot(date) ...

}
\keyword{datasets}
