% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{date_type}
\alias{date_type}
\title{Type of dates in date.rda}
\format{
A data frame with 31 observations on the following 4 variables.
\describe{ \item{ID_DATE_TYPE}{a numeric vector}
\item{DATE_TYPE_CODE}{a character vector}
\item{DATE_TYPE_NAME}{a character vector}
\item{DATE_TYPE_NUMBER}{a numeric vector} }
obtained ~~
}
\description{
Type of dates in date.rda
}
\examples{

data(date_type)
## maybe str(date_type) ; plot(date_type) ...

}
\keyword{datasets}
