% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{depo_context}
\alias{depo_context}
\title{Depositional contexts}
\format{
A data frame with 17 observations on the following 4 variables.
\describe{ \item{ID_DEPO_CONTEXT}{a numeric vector}
\item{DEPO_CONTEXT_CODE}{a character vector}
\item{DEPO_CONTEXT_NAME}{a character vector}
\item{DEPO_CONTEXT_NUMBER}{a numeric vector} }
}
\description{
Depositional contexts
}
\examples{

data(depo_context)
## maybe str(depo_context) ; plot(depo_context) ...

}
\keyword{datasets}
