% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_slc.R
\name{dist_slc}
\alias{dist_slc}
\title{Calculates distance between geocoordinates: Spherical Law of Cosines (SLC)}
\usage{
dist_slc(lat1, lon1, lat2, lon2, type = "deg", km = TRUE)
}
\arguments{
\item{lat1}{the latitude as radians of the first point}

\item{lon1}{the longitude as radians of the first point}

\item{lat2}{the latitude as radians of the second point}

\item{lon2}{the longitude as radians of the second point}

\item{type}{defaults to "deg", can also be "rad"}

\item{km}{boolean argument for whether to return results as km (TRUE) or
miles (FALSE)}
}
\description{
\code{dist_slc} takes inputs of two sets of coordinates in
(radian values), one set fo reach location, and a boolean indicator of
whether or not to return the results as kilometers (\code{km = TRUE}) or
miles (\code{km = FALSE}). The output is the distance using the method
of Spherical Law of Cosines, a basic distance between two points on the
surface of a sphere.
}
\examples{
# Input list of degree values
# Longitude values range between 0 and +-180 degrees
deg.lon <- runif(2, -180, 180)
# Latitude values range between 0 and +-90 degrees
deg.lat <- runif(2, -90, 90)

# Obtain measures of distnace
sphere.mi <- dist_slc(lon1 = deg.lon[1], lat1 = deg.lat[1]
  , lon2 = deg.lon[2], lat2 = deg.lat[2], km = FALSE)
}
