% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{calcualte the Euclidean Distance of two number}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{number 1}

\item{b}{number 2}
}
\value{
return the greatest common divisor of number 1 and number 2
}
\description{
calcualte the Euclidean Distance of two number
}
\examples{
euclidean(4,12)
}
