% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{package}
\name{GCD-package}
\alias{GCD-package}
\alias{GCD}
\title{GCD: Global Charcoal Database data}
\description{
The GCD package provides Global Charcoal Database data updated at regular
intervals to be used in cunjuction with the paleofire package.
}
\details{
\tabular{ll}{ Package: \tab GCD\cr Type: \tab Package\cr Version: \tab
4.0.0\cr Date: \tab 2018-03-01\cr License: \tab GPL (>=2)\cr }
}
\section{Author(s)}{
 Global Paleofire Working Group <paleofire gmail.com>
}

\examples{

rm(list=ls())

library(GCD)

## Charcoal data:

data(paleofiredata)
head(paleofiredata)


## Associated sites informations:

data(paleofiresites)
head(paleofiresites)

}
\references{
Power, M., J. Marlon, P. Bartlein, and S. Harrison. 2010. Fire
history and the Global Charcoal Database: a new tool for hypothesis testing
and data exploration. Palaeogeography, Palaeoclimatology, Palaeoecology
291:52-59.
}
\seealso{
\url{http://paleofire.org}
}
\concept{charcoal}
\concept{fire}
\concept{global}
\concept{paleo}
\concept{sediments}
