% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.HERE_geocode.R
\name{gcd.HERE_geocode}
\alias{gcd.HERE_geocode}
\title{Get geocoding information using HERE.com Geocoder API}
\usage{
gcd.HERE_geocode(address, app_id, app_code, dev = FALSE, verbose = FALSE)
}
\arguments{
\item{address}{the latitude coordinate for the origin location}

\item{app_id}{the longitude coordinate for the origin location}

\item{app_code}{the latitude coordinate for the destination location}

\item{dev}{whether to use development or production site}

\item{verbose}{Logical argument determining if messages are displayed.}
}
\value{
location information for the entered address string:
  OrigAddr, Latitude, Longitude, Label, Country, State, County, City,
  District, Street, HouseNumber, PostalCode
}
\description{
The function \code{"gcd.HERE_geocode"} is designed to take an address
string, and user app credentials. The function queries the HERE.com Geocoder
API
\code{(https://developer.here.com/documentation/geocoder/topics/what-is.html)}
and returns the drive time from the JSON output code. This geocoding
information from the JSON is originally returned as a dataframe.
}
