% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.rad.R
\name{gcd.rad}
\alias{gcd.rad}
\title{Converts degrees to radians}
\usage{
gcd.rad(deg)
}
\arguments{
\item{deg}{a value in degrees}
}
\description{
The function \code{"gcd.rad"} takes an input of a degree value and will
convert it to a radian value by the formula:
\code{"radian = degree * pi / 180"}
Typically, this conversion would be in preparation of geocoordinate
calculations in which the desired form of the coordinates is radians.
}
