% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\docType{package}
\name{GCD}
\alias{GCD}
\alias{GCD-package}
\title{GCD: A package for computing the GCD of two numbers}
\description{
The GCD package contains one function, the euclidean function.
}
\section{euclidean functions}{

The \link[GCD]{euclidean} function takes GCD of two numbers using euclidean algortihm.
}

