% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.slc.R
\name{gcd.slc}
\alias{gcd.slc}
\title{Calculates the distance between geocoordinates using the Spherical Law of
Cosines (SLC)}
\usage{
gcd.slc(lon1, lat1, lon2, lat2, km = TRUE)
}
\arguments{
\item{lon1}{the longitude as radians of the first point}

\item{lat1}{the latitude as radians of the first point}

\item{lon2}{the longitude as radians of the second point}

\item{lat2}{the latitude as radians of the second point}

\item{km}{boolean argument for whether to return results as km (TRUE) or
miles (FALSE)}
}
\description{
The function \code{"gcd.slc"} takes inputs of two sets of coordinates in
(radian values), one set fo reach location, and a boolean indicator of
whether or not to return the results as kilometers (\code{"km = TRUE"}) or
miles (\code{"km = FALSE"}). The output is the distance using the method
of Spherical Law of Cosines, a basic distance between two points on the
surface of a sphere.
}
