% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.vincenty.R
\name{gcd.vincenty}
\alias{gcd.vincenty}
\title{Calculates the distance between geocoordinates using the Vincenty inverse
formula for ellipsoids}
\usage{
gcd.vincenty(lon1, lat1, lon2, lat2, km = TRUE)
}
\arguments{
\item{lon1}{the longitude as radians of the first point}

\item{lat1}{the latitude as radians of the first point}

\item{lon2}{the longitude as radians of the second point}

\item{lat2}{the latitude as radians of the second point}

\item{km}{boolean argument for whether to return results as km (TRUE) or
miles (FALSE)}
}
\description{
The function \code{"gcd.vincenty"} takes inputs of two sets of coordinates
in (radian values), one set fo reach location, and a boolean indicator of
whether or not to return the results as kilometers (\code{"km = TRUE"}) or
miles (\code{"km = FALSE"}). The output is the distance using the method
of the Vincenty inverse formula for ellipsoinds, a more robust distance
between two points on the surface of an ellipsoid (Earth).
}
