% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Google_drivetime.R
\name{Google_drivetime}
\alias{Google_drivetime}
\title{Get drive time in hours from Google.com Routing API}
\usage{
Google_drivetime(origin = NULL, destination = NULL, key = NULL,
  mode = "driving", transit_mode = NULL,
  transit_routing_preference = NULL, time_format = "minutes",
  distance_format = "customary", coordinate_type = "deg",
  verbose = FALSE)
}
\arguments{
\item{origin}{a string or vector of two numerics; if a location string,
this should be an address, if a vector of length two numerics this should
be the latitude and longitude coordinates for the origin location}

\item{destination}{a string or vector of two numerics; if a location string,
this should be an address, if a vector of length two numerics this should
be the latitude and longitude coordinates for the destination location}

\item{key}{a string; the user's App Code for the Google.com JavaScript/REST
(requires registration)}

\item{mode}{a string; a transportation method for the route:
\enumerate{
\item \code{"driving"}
\item \code{"walking"}
\item \code{"bicycling"}
\item \code{"transit"}
}}

\item{transit_mode}{a string; sub-type for \code{"transit"} otherwise
should be NULL:
\enumerate{
\item \code{"bus"}: the calculated route should prefer travel by bus
\item \code{"subway"}: the calculated route should prefer travel by subway
\item \code{"train"}: the calculated route should prefer travel by train
\item \code{"tram"}: the calculated route should prefer travel by tram and
light rail
\item \code{"rail"}: the calculated route should prefer travel by train,
tram, light rail, and subway
}}

\item{transit_routing_preference}{a string; specifies preferences for the
\code{"transit"}, otherwise should be NULL:
\enumerate{
\item \code{"less_walking"}
\item \code{"fewer_transfers"}
}}

\item{time_format}{a string; the preferred time format for results:
\enumerate{
\item \code{"hours"}
\item \code{"minutes"}
\item \code{"seconds"}
}}

\item{distance_format}{a string; the pregerred format for measuring
distance:
\enumerate{
\item \code{"customary"}
\item \code{"metric"}
}}

\item{coordinate_type}{a string; if using geocoordinates, whether they are
in degrees or radians:
\enumerate{
\item \code{"rad"}{ (radians)}
\item \code{"deg"}{ (degrees)}
}}

\item{verbose}{a boolean; \code{TRUE} to display json webaddress,
\code{FALSE} to not display json webaddress}
}
\value{
the travel time and distance as specified between two locations
}
\description{
\code{Google_drivetime} is designed to take degree or
radian geocoordinates or location names, user app credentials
, and some other API inputs (set to defaults). The function queries the
Google Directions API \code{(
https://developers.google.com/maps/documentation/directions/
intro#DirectionsRequests)} and return
the drive time from the JSON output code. This drive time portion of the JSON
is originally return in minutes. This function is designed to take an
argument for the user preferred time (seconds, minutes, or hours).
}
