% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Google_geocode.R
\name{Google_geocode}
\alias{Google_geocode}
\title{Get geocoding information using Google Geocode API}
\usage{
Google_geocode(address = NULL, key = NULL, verbose = FALSE)
}
\arguments{
\item{address}{an address as a string}

\item{key}{the user's Google API key}

\item{verbose}{Logical argument determining if messages are displayed.}
}
\value{
location information for the entered address string:
  original_address, latitude, longitude, formatted_address, street_number
  , route, locality (city), administrative_area_3 (district)
  , administrative_area_2 (county), administrative_area_1 (state)
  , country, postal_code
}
\description{
\code{Google_geocode} is designed to take an address
  string, and user app key. The function queries the Google Geocode
  API
\code{
  (https://developers.google.com/maps/documentation/geocoding/start)
  }
  and returns the various components from the JSON output code. This geocoding
  information from the JSON is originally returned as a vector.
}
