% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERE_geocode.R
\name{HERE_geocode}
\alias{HERE_geocode}
\title{Get geocoding information using HERE.com Geocoder API}
\usage{
HERE_geocode(address = NULL, app_id = NULL, app_code = NULL,
  dev = FALSE, verbose = FALSE)
}
\arguments{
\item{address}{an address as a stringn}

\item{app_id}{the user's App ID for the HERE.com JavaScript/REST (requires
registration)}

\item{app_code}{the user's App Code for the HERE.com JavaScript/REST
(requires registration)}

\item{dev}{whether to use development or production site}

\item{verbose}{Logical argument determining if messages are displayed.}
}
\value{
location information for the entered address string:
  OrigAddr, Latitude, Longitude, Label, Country, State, County, City,
  District, Street, HouseNumber, PostalCode
}
\description{
\code{HERE_geocode} is designed to take an address
string, and user app credentials. The function queries the HERE.com Geocoder
API
\code{
  (https://developer.here.com/documentation/geocoder/topics/what-is.html)
  }
  and returns the drive time from the JSON output code. This geocoding
  information from the JSON is originally returned as a dataframe.
}
