% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERE_isoline.R
\name{HERE_isoline}
\alias{HERE_isoline}
\title{Get polygon line drawing output from HERE.com}
\usage{
HERE_isoline(app_id = NULL, app_code = NULL, origin = NULL,
  coord_type = "deg", mode_type = "fastest", mode_tran = "car",
  mode_traf = "disabled", range_type = "distance", range_unit = "mi",
  range_val = NULL, dev = FALSE, verbose = FALSE)
}
\arguments{
\item{app_id}{a string; the user's App ID for the HERE.com JavaScript/REST
(requires registration)}

\item{app_code}{a string; the user's App Code for the HERE.com
JavaScript/REST (requires registration)}

\item{origin}{a vector of length two (2); the latitude and longitude
coordinates of the origin location}

\item{coord_type}{a string; if the geocoordinates are in degrees or radians:
\enumerate{
\item \code{"rad"}{ (radians)}
\item \code{"deg"}{ (degrees)}
}}

\item{mode_type}{a string; a method of deciding optimal travel route:
\enumerate{
\item \code{"fastest"}
\item \code{"shortest"}
}}

\item{mode_tran}{a string; a transportation method for the route:
\enumerate{
\item \code{"car"}
\item \code{"carHOV"}
\item \code{"pedestrian"}
\item \code{"truck"}
\item \code{"bicycle"}
}}

\item{mode_traf}{a string; whether or not to factor in traffic:
\enumerate{
\item \code{"disabled"}
\item \code{"enabled"}
}}

\item{range_type}{a string; what range constraint the user desires to use:
\enumerate{
\item \code{"distance"}
\item \code{"time"}
}}

\item{range_unit}{a string; the units the provided range are given in:
\enumerate{
\item \code{"mi"}
\item \code{"ft"}
\item \code{"km"}
\item \code{"m"}
\item \code{"hrs"}
\item \code{"min"}
\item \code{"sec"}
}}

\item{range_val}{a numeric; the value of the range constraint}

\item{dev}{a boolean; whether to use development or production site}

\item{verbose}{a boolean; \code{TRUE} to display json webaddress,
\code{FALSE} to not display json webaddress}
}
\value{
the polygon data.frame
}
\description{
\code{HERE_isoline} is designed to take
}
