% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{lands_desc}
\alias{lands_desc}
\title{Landscape description elements}
\format{
A data frame with 20 observations on the following 3 variables.
\describe{ \item{ID_LANDS_DESC}{a numeric vector}
\item{LANDS_DESC_CODE}{a character vector}
\item{LANDS_DESC_NAME}{a character vector} }
}
\description{
Landscape description elements
}
\examples{

data(lands_desc)
## maybe str(lands_desc) ; plot(lands_desc) ...

}
\keyword{datasets}
