% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{lcm}
\alias{lcm}
\title{Least Common Multiple}
\usage{
lcm(a, b)
}
\arguments{
\item{a}{Positive integer.}

\item{b}{Positive integer.}
}
\description{
Calculates the least common multiple of two positive integers.
}
\examples{
lcm(45, 18)
}
\seealso{
\code{\link[=gcd]{gcd()}}
}
