% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{mat_dated}
\alias{mat_dated}
\title{Material Dated}
\format{
A data frame with 23 observations on the following 4 variables.
\describe{ \item{ID_MAT_DATED}{a numeric vector}
\item{MAT_DATED_STANDARD_LEVEL}{a character vector}
\item{MAT_DATED_TYPE}{a character vector}
\item{MAT_DATED_HIGH_LEVEL}{a character vector} }
obtained ~~
}
\description{
Material Dated
}
\examples{

data(mat_dated)
## maybe str(mat_dated) ; plot(mat_dated) ...

}
\keyword{datasets}
