% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{method}
\alias{method}
\title{Methods}
\format{A data frame with 13 observations on the following 3 variables.
\describe{ \item{list("id_method")}{a factor with levels \code{ACID}
\code{CPRO} \code{GRAV} \code{HNPK} \code{HVLQ} \code{IMAG} \code{NOTK}
\code{OPPF} \code{OTHE} \code{POLS} \code{SIEV} \code{SIPO} \code{THSL}}
\item{list("method")}{a factor with levels \code{Acidification of sample
using gas chromatography to measure elemental carbon (Verardo etal 1990)}
\code{charcoal part Ided as opaque phytoclasts/pyrofusinite in palynofacies
anlaysis} \code{Charcoal part. Ided in thin slides (soil micromorphology)}
\code{Charcoal particles identified by imaging software} \code{Charcoal
separated by heavy liquid preparation} \code{Cumulative probability (95%
confidence interval) alluvial soil charcoal} \code{GRAVIMETRIC chemical
assay (Winkler method)} \code{Hand picked charcoal from soil samples}
\code{Not known} \code{Other} \code{Pollen slide} \code{Sieved} \code{Sieved
+ Pollslide}} \item{list("code")}{a numeric vector} }}
\description{
Charcoal methods used in GCD
}
\examples{

data(method)

}
\keyword{datasets}

