% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ny_sport-data.R
\docType{data}
\name{ny_sports}
\alias{ny_sports}
\title{Sports Teams of the NY City Area}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(ny_sports)
}
\description{
Data for the NY City Area sports teams with the team name, sport,
performance location name, and performance location address.
}
\examples{
data(ny_sports)
ny_sports
}
\keyword{datasets}
