% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-package.R
\docType{data}
\name{site_type}
\alias{site_type}
\title{Type of sites}
\format{
A data frame with 53 observations on the following 4 variables.
\describe{ \item{ID_SITE_TYPE}{a numeric vector}
\item{SITE_TYPE_LEVEL}{a character vector}
\item{SITE_TYPE_DESC}{a character vector}
\item{SITE_TYPE_HIGH_LEVEL}{a character vector} }
}
\description{
Type of sites
}
\examples{

data(site_type)
## maybe str(site_type) ; plot(site_type) ...

}
\keyword{datasets}
