% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deg.R
\name{to_deg}
\alias{to_deg}
\title{Converts radians to degrees}
\usage{
to_deg(rad)
}
\arguments{
\item{rad}{a value in radians}
}
\description{
\code{to_deg} takes an input of a radian value and will
convert it to a degree value by the formula:
\code{degree = radian * 180 / pi}
This utility conversion is useful when dealing with geocoding coordinates,
as it might be desirable to obtain the original coordinate values in
degrees.
}
\examples{
# Input list of degree values
# Longitude values range between 0 and +-180 degrees
rad <- runif(1000, -180, 180) * pi / 180
deg <- to_deg(rad)
}
