% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rad.R
\name{to_rad}
\alias{to_rad}
\title{Converts degrees to radians}
\usage{
to_rad(deg)
}
\arguments{
\item{deg}{a value in degrees}
}
\description{
\code{to_rad} takes an input of a degree value and will
convert it to a radian value by the formula:
\code{radian = degree * pi / 180}
Typically, this conversion would be in preparation of geocoordinate
calculations in which the desired form of the coordinates is radians.
}
\examples{
# Input list of degree values
# Longitude values range between 0 and +-180 degrees
deg <- runif(1000, -180, 180)
rad <- to_rad(deg)
}
