% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest Common Divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is a numeric scalar or integer.}

\item{b}{is a numeric scalar or integer.}
}
\value{
The greates common divisor of a and b.
}
\description{
Takes the GCD of two numbers by using Euclidean algortihm. See \href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia}
}

