% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest Common Divisor}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{Positive integer.}

\item{b}{Positive integer.}
}
\value{
Positive integer.
}
\description{
Calculates the greatest common divisor of two positive integers.
}
\examples{
gcd(6, 8)
gcd(234987, 83268113424)
}
\seealso{
\code{\link[=lcm]{lcm()}}
}
