% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Calculates the distance between geocoordinates using the Spherical Law of
Cosines, Haversine formula, and Vincenty inverse formula for ellipsoids.}
\usage{
gcd(lon1, lat1, lon2, lat2, km = TRUE)
}
\arguments{
\item{lon1}{the longitude as radians of the first point}

\item{lat1}{the latitude as radians of the first point}

\item{lon2}{the longitude as radians of the second point}

\item{lat2}{the latitude as radians of the second point}

\item{km}{boolean argument for whether to return results as km (TRUE) or
miles (FALSE)}
}
\description{
The function \code{"gcd"} takes inputs of two sets of coordinates
in (degree values), one set fo reach location, and a boolean indicator of
whether or not to return the results as kilometers (\code{"km = TRUE"}) or
miles (\code{"km = FALSE"}). The output is three distances using the methods
of Spherical Law of Cosines, Haversine formula, and Vincenty inverse formula
for ellipsoids.
}
