% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERE_drivetime.R
\name{HERE_drivetime}
\alias{HERE_drivetime}
\title{Get drive time in hours from HERE.com Routing API}
\usage{
HERE_drivetime(orgn_lat, orgn_lon, dest_lat, dest_lon, app_id, app_code,
  time_frmt = "hours", type = "fastest", trnsprt = "car",
  trfc = "disabled", coord_typ = "deg", dev = FALSE,
  verbose = FALSE)
}
\arguments{
\item{orgn_lat}{the latitude coordinate for the origin location}

\item{orgn_lon}{the longitude coordinate for the origin location}

\item{dest_lat}{the latitude coordinate for the destination location}

\item{dest_lon}{the longitude coordinate for the destination location}

\item{app_id}{the user's App ID for the HERE.com JavaScript/REST (requires
registration)}

\item{app_code}{the user's App Code for the HERE.com JavaScript/REST
(requires registration)}

\item{time_frmt}{the user's preferred time format: \enumerate{ \item
\code{"hours"} \item \code{"minutes"} \item \code{"seconds"} }}

\item{type}{a method of deciding optimal travel route: \enumerate{ \item
\code{"fastest"} \item \code{"shortest"} }}

\item{trnsprt}{a transportation method for the route: \enumerate{ \item
\code{"car"} \item \code{"carHOV"} \item \code{"pedestrian"} \item
\code{"truck"} \item \code{"bicycle"} }}

\item{trfc}{whether or not to factor in traffic: \enumerate{ \item
\code{"disabled"} \item \code{"enabled"} }}

\item{coord_typ}{if the geocoordinates are in degrees or radians:
\enumerate{ \item \code{"rad"}{ (radians)} \item \code{"deg"}{ (degrees)} }}

\item{dev}{whether to use development or production site}

\item{verbose}{Logical argument determining if messages are displayed.}
}
\value{
the travel time as specified between two locations
}
\description{
\code{HERE_drivetime} is designed to take degree or
radian geocoordinates, user app credentials, and some other API inputs (set
to defaults). The function queries the HERE.com Routing API \code{(https://
developer.here.com/documentation/routing/topics/what-is.html)} and return
the drive time from the JSON output code. This driveTime portion of the JSON
is originally return in seconds. This function is designed to take an
argument for the user preferred time (seconds, minutes, or hours) as a non-
integer, numeric value (a number with decimals).
}
