% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HERE_geocode.R
\name{HERE_geocode}
\alias{HERE_geocode}
\title{Get geocoding information using HERE.com Geocoder API}
\usage{
HERE_geocode(
  address = NULL,
  app_api = NULL,
  app_id = NULL,
  app_code = NULL,
  dev = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{address}{a string; the address to be queried for geocoded information.}

\item{app_api}{a string; the user's API Key Credentials for the HERE.com
JavaScript/REST (requires registration)}

\item{app_id}{a string; the user's App ID for the HERE.com JavaScript/REST
(requires registration)}

\item{app_code}{a string; the user's App Code for the HERE.com
JavaScript/REST (requires registration)}

\item{dev}{a boolean; \code{TRUE} to use the development app, \code{FALSE}
to use the production app}

\item{verbose}{a boolean; \code{TRUE} to display json webaddress,
\code{FALSE} to not display json webaddress}
}
\value{
location information for the entered address string:
  OrigAddr, Latitude, Longitude, Label, Country, State, County, City,
  District, Street, HouseNumber, PostalCode
}
\description{
\code{HERE_geocode} is designed to take an address
string, and user app credentials. The function queries the HERE.com Geocoder
API
\code{
  (https://developer.here.com/documentation/geocoder/topics/what-is.html)
  }
  and returns the drive time from the JSON output code. This geocoding
  information from the JSON is originally returned as a dataframe.
}
