\name{paleofiredata}
\alias{paleofiredata}
\docType{data}
\title{
GCD Charcoal data}
\description{
Charcoal series from the Global harcoal Database}
\usage{data(paleofiredata)}
\format{
  A data frame with 118832 observations on the following 4 variables.
  \describe{
    \item{\code{ID_SITE}}{a numeric vector}
    \item{\code{DEPTH}}{a numeric vector}
    \item{\code{EST_AGE}}{a numeric vector}
    \item{\code{QUANTITY}}{a numeric vector}
  }
}
\examples{
data(paleofiredata)
## maybe str(paleofiredata) ; plot(paleofiredata) ...
}
\keyword{datasets}
